requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.6.2
    # via flask
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   flask
    #   functions-framework
    #   google-ssl (pyproject.toml)
cloudevents==1.9.0
    # via functions-framework
deprecation==2.1.0
    # via cloudevents
flask==2.3.2
    # via functions-framework
functions-framework==3.3.0
    # via google-ssl (pyproject.toml)
google-api-core[grpc]==2.11.0
    # via
    #   google-api-python-client
    #   google-cloud-compute
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.86.0
    # via google-ssl (pyproject.toml)
google-auth==2.17.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-compute==1.11.0
    # via google-ssl (pyproject.toml)
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-storage==2.8.0
    # via google-ssl (pyproject.toml)
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-storage
googleapis-common-protos==1.59.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.54.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.54.0
    # via google-api-core
gunicorn==20.1.0
    # via functions-framework
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
packaging==23.1
    # via deprecation
proto-plus==1.22.2
    # via google-cloud-compute
protobuf==4.22.3
    # via
    #   google-api-core
    #   google-cloud-compute
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.0.9
    # via httplib2
requests==2.29.0
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.15
    # via requests
watchdog==3.0.0
    # via functions-framework
werkzeug==2.3.3
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
